Applied the patch for STANBOL-1317 with two modifications reflection recent changes in the 0.12 branch
git-svn-id: https://svn.apache.org/repos/asf/stanbol/branches/release-0.12@1587559 13f79535-47bb-0310-9956-ffa450edef68

+++ commons/indexedgraph/src/test/java/org/apache/stanbol/commons/indexedgraph/IndexedGraphTest.java
@@ -492,7 +492,7 @@ public class IndexedGraphTest  extends MGraphTest {
                 tc.add(new TripleImpl(subject, predicate, bnode));
             } else { //UriRef
                 tc.add(new TripleImpl(subject, predicate,
-                        new UriRef(URI_PREFIX + (int) count * random)));
+                        new UriRef(URI_PREFIX + count * random)));
             }
         }        
     }
