+++ commons/solr/core/src/main/java/org/apache/stanbol/commons/solr/utils/ServiceReferenceRankingComparator.java
@@ -36,12 +36,12 @@ public class ServiceReferenceRankingComparator implements Comparator<ServiceRefe
         int r1,r2;
         Object tmp = ref1.getProperty(Constants.SERVICE_RANKING);
         r1 = tmp != null ? ((Integer)tmp).intValue() : 0;
-        tmp = (Integer)ref2.getProperty(Constants.SERVICE_RANKING);
+        tmp = ref2.getProperty(Constants.SERVICE_RANKING);
         r2 = tmp != null ? ((Integer)tmp).intValue() : 0;
         if(r1 == r2){
-            tmp = (Long)ref1.getProperty(Constants.SERVICE_ID);
+            tmp = ref1.getProperty(Constants.SERVICE_ID);
             long id1 = tmp != null ? ((Long)tmp).longValue() : Long.MAX_VALUE;
-            tmp = (Long)ref2.getProperty(Constants.SERVICE_ID);
+            tmp = ref2.getProperty(Constants.SERVICE_ID);
             long id2 = tmp != null ? ((Long)tmp).longValue() : Long.MAX_VALUE;
             //the lowest id must be first -> id1 < id2 -> [id1,id2] -> return -1
             return id1 < id2 ? -1 : id2 == id1 ? 0 : 1;
